# api/openapi.yaml
openapi: 3.1.0
info:
  title: blabby HTTP API
  version: 0.1.0
  description: |
    The HTTP API for blabby, a real-time chat service.

    Clients use HTTP for commands and queries (login, room discovery, membership,
    and sending messages) and a WebSocket for the live event stream. The WebSocket
    message contract is documented separately in `asyncapi.yaml`.

    ## Authentication

    Every endpoint requires an `Authorization: Bearer <jwt>` header **except**:

    - `POST /users`, `POST /users/verifications`, and
      `POST /users/verifications/resend` — the pre-login registration flow,
    - `POST /login`, which mints the token and is therefore public, and
    - `GET /ws`, where authentication is the first WebSocket frame after the
      upgrade rather than an HTTP header.

    A missing or malformed bearer header is rejected with `401 AUTH_MISSING_TOKEN`,
    an expired token with `401 AUTH_EXPIRED_TOKEN`, and any other invalid token
    with `401 AUTH_INVALID_TOKEN`.

    ## Error envelope

    Every application-generated JSON error response shares one envelope: a top-level
    `error` object holding a numeric `code`, a canonical `status` string, and a
    human-readable, client-safe `message`. A failed WebSocket upgrade is emitted by
    the WebSocket library as plain text instead. The `status` enum on `ErrorDetail`
    lists the full error taxonomy; each operation's `responses` section is the
    authoritative list of what that operation can actually return.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /users/verifications:
    post:
      operationId: verifyRegistration
      tags:
        - Registration
      summary: Activate a pending account with its emailed PIN.
      description: |
        Confirms a pending account by checking the submitted PIN against its
        challenge. On success the account becomes `active` — it can now log in —
        and the challenge is cleared. This endpoint is public.

        Failure is deliberately uniform: an unknown or non-pending address, a
        missing or expired challenge, a wrong PIN, and a challenge locked by too
        many wrong attempts (5) all return the same `400 VERIFICATION_INVALID`,
        so the response reveals nothing about which precondition failed. A
        locked challenge is recovered by requesting a fresh PIN via
        `POST /users/verifications/resend`, which resets the attempt count.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
            example:
              mail_address: dana@example.com
              pin: '482910'
      responses:
        '200':
          description: The account is now active and can log in. The response has no body.
        '400':
          description: |
            The request was malformed or not `application/json`
            (`INVALID_REQUEST`), the email did not parse (`INVALID_EMAIL`), or
            the verification could not succeed (`VERIFICATION_INVALID` — the
            uniform failure described above).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 6003
                  status: VERIFICATION_INVALID
                  message: verification failed
        '413':
          description: The request body exceeded the 1 MB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4003
                  status: PAYLOAD_TOO_LARGE
                  message: request body exceeds maximum size
        '500':
          $ref: '#/components/responses/InternalError'
  /users/verifications/resend:
    post:
      operationId: resendVerificationPIN
      tags:
        - Registration
      summary: Re-issue a verification PIN to a pending account.
      description: |
        Sends a fresh PIN to a pending account, replacing the previous challenge
        and resetting its attempt count. This endpoint is public.

        To avoid revealing which addresses are registered, an unknown or
        already-active address is a silent no-op that still returns `200`. Only
        a pending account that has exhausted its resend budget returns `429`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
            example:
              mail_address: dana@example.com
      responses:
        '200':
          description: |
            A fresh PIN was sent — or the address was unknown or already active
            and nothing happened; the two outcomes are indistinguishable by
            design. The response has no body.
        '400':
          description: |
            The request body was malformed or not `application/json`
            (`INVALID_REQUEST`), or the email did not parse (`INVALID_EMAIL`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4004
                  status: INVALID_EMAIL
                  message: mail address is not a valid email
        '413':
          description: The request body exceeded the 1 MB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4003
                  status: PAYLOAD_TOO_LARGE
                  message: request body exceeds maximum size
        '429':
          $ref: '#/components/responses/VerificationRateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /rooms/{id}/members/{user}/role:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/UserCode'
    put:
      operationId: setRoomMemberRole
      tags:
        - Rooms
      summary: Set another member's role.
      description: |
        Changes the target member's role to `admin` or `member`. The caller must
        be a member whose role permits it — the owner or an admin; admins may
        promote members and demote admins alike. The owner role never moves
        through this endpoint: replacing the owner is `PUT /rooms/{id}/owner`.
        Setting a role the member already holds succeeds, so the PUT is safe to
        repeat.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRoleRequest'
            example:
              role: admin
      responses:
        '200':
          description: The member now holds the requested role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
        '400':
          description: |
            The room or user code was malformed, the user code matches no
            account, the body was malformed or missing `role`, or the role was
            not `admin`/`member`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4001
                  status: INVALID_REQUEST
                  message: role must be admin or member
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            The caller or the target is not a member (`ROOM_NOT_MEMBER`), or the
            caller's role does not permit the change (`ROOM_PERMISSION_DENIED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 2005
                  status: ROOM_PERMISSION_DENIED
                  message: your role does not permit changing member roles
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '413':
          description: The request body exceeded the 1 KiB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4003
                  status: PAYLOAD_TOO_LARGE
                  message: request body exceeds maximum size
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /rooms/{id}/owner:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    put:
      operationId: transferRoomOwnership
      tags:
        - Rooms
      summary: Transfer the room's ownership.
      description: |
        Hands the room to another member: the caller (who must be the current
        owner) is demoted to admin and the named member becomes the owner. The
        owner is modelled as a REST resource, so naming the current owner is an
        idempotent `200` no-op. An owner cannot leave a room
        (`409 ROOM_OWNER_CANNOT_LEAVE` on the membership DELETE) until they have
        transferred it here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetOwnerRequest'
            example:
              user: UB000000002
      responses:
        '200':
          description: The named member owns the room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
        '400':
          description: |
            The room code was malformed, the body was malformed or missing
            `user`, or the user code was malformed or matches no account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4001
                  status: INVALID_REQUEST
                  message: unknown user
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            The caller or the named member is not a member (`ROOM_NOT_MEMBER`),
            or the caller is not the owner (`ROOM_PERMISSION_DENIED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 2005
                  status: ROOM_PERMISSION_DENIED
                  message: only the owner can transfer ownership
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '413':
          description: The request body exceeded the 1 KiB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4003
                  status: PAYLOAD_TOO_LARGE
                  message: request body exceeds maximum size
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "A JWT issued by `POST /login`, sent as `Authorization: Bearer <jwt>`."
  parameters:
    RoomId:
      name: id
      in: path
      required: true
      description: |
        The opaque public room code: `R` followed by 10 Crockford Base32 symbols
        (e.g. `RG000000004`). Input is parsed leniently (case-insensitive, `O`→`0`,
        `I`/`L`→`1`). A malformed code is rejected with `400 INVALID_REQUEST`; a
        well-formed code that matches no active room is `404 ROOM_NOT_FOUND`. The
        internal numeric room id is never accepted here.
      schema:
        type: string
      example: RG000000004
    UserCode:
      name: user
      in: path
      required: true
      description: |
        The target member's opaque public user code: `U` followed by 10
        Crockford Base32 symbols (e.g. `UB000000002`). Parsed with the same
        leniency as room codes. A malformed or unknown code is rejected with
        `400 INVALID_REQUEST` — the codes are unguessable, and a nonexistent
        user is by definition not a member, so nothing is revealed that
        membership wouldn't.
      schema:
        type: string
      example: UB000000002
  schemas:
...
    VerifyRequest:
      type: object
      required:
        - mail_address
        - pin
      properties:
        mail_address:
          type: string
          format: email
          x-max-utf8-bytes: 254
          description: The email address the account registered with.
        pin:
          type: string
          pattern: '^[0-9]{6}$'
          description: The 6-digit PIN from the verification email.
    ResendVerificationRequest:
      type: object
      required:
        - mail_address
      properties:
        mail_address:
          type: string
          format: email
          x-max-utf8-bytes: 254
          description: The email address the account registered with.
...
    SetRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - admin
            - member
          description: |
            The target member's new role. The owner role is not assignable here;
            it moves only through `PUT /rooms/{id}/owner`.
    SetOwnerRequest:
      type: object
      required:
        - user
      properties:
        user:
          type: string
          description: |
            The new owner's opaque public user code (`U…`). Must be a member of
            the room.
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Always `true` on a 200 response.
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - status
        - message
      properties:
        code:
          type: integer
          description: The numeric error code paired with `status`.
        status:
          type: string
          description: |
            The canonical status string. This enum is the complete error
            taxonomy, not the subset any single endpoint returns; consult each
            operation's `responses` for the codes it can actually produce.
          enum:
            - AUTH_INVALID_TOKEN
            - AUTH_EXPIRED_TOKEN
            - AUTH_MISSING_TOKEN
            - AUTH_ACCOUNT_PENDING
            - ROOM_NOT_MEMBER
            - ROOM_ALREADY_MEMBER
            - ROOM_NOT_FOUND
            - ROOM_OWNER_CANNOT_LEAVE
            - ROOM_PERMISSION_DENIED
            - RATE_LIMIT_EXCEEDED
            - VERIFICATION_RATE_LIMITED
            - INVALID_REQUEST
            - MISSING_FIELD
            - PAYLOAD_TOO_LARGE
            - INVALID_EMAIL
            - WEAK_PASSWORD
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - EMAIL_ALREADY_REGISTERED
            - HANDLE_ALREADY_TAKEN
            - VERIFICATION_INVALID
        message:
          type: string
          description: A human-readable, client-safe explanation. Never leaks internal detail.
  responses:
    VerificationRateLimited:
      description: |
        The verification-PIN send budget is exhausted: a PIN was sent too
        recently (minimum interval) or the account has hit its resend cap. Wait
        before retrying.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 3002
              status: VERIFICATION_RATE_LIMITED
              message: too many verification attempts; please wait and try again
    Unauthorized:
      description: |
        The bearer token is missing, malformed, expired, or otherwise invalid
        (`AUTH_MISSING_TOKEN`, `AUTH_EXPIRED_TOKEN`, or `AUTH_INVALID_TOKEN`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 1003
              status: AUTH_MISSING_TOKEN
              message: missing authorization header
...
    RoomNotFound:
      description: |
        The room code is well-formed but matches no active room (the room does not
        exist or is no longer active).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 2003
              status: ROOM_NOT_FOUND
              message: room not found
    ServiceUnavailable:
      description: A backend dependency was temporarily unreachable. The request may be retried.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 5002
              status: SERVICE_UNAVAILABLE
              message: failed to reach user grain
    InternalError:
      description: An unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 5001
              status: INTERNAL_ERROR
              message: authentication unavailable